+++ builtin/mainmenu/pkgmgr.lua
@@ -547,7 +547,9 @@ function pkgmgr.install_dir(type, path, basename, targetpath)
 		if not targetpath then
 			targetpath = core.get_texturepath() .. DIR_DELIM .. basename
 		end
-		core.delete_dir(targetpath)
+		if core.is_dir(targetpath) then
+			core.delete_dir(targetpath)
+		end
 		if not core.copy_dir(from, targetpath, false) then
 			return nil,
 				fgettext("Failed to install $1 to $2", basename, targetpath)
@@ -567,9 +569,7 @@ function pkgmgr.install_dir(type, path, basename, targetpath)
 		end
 
 		-- Get destination name for modpack
-		if targetpath then
-			core.delete_dir(targetpath)
-		else
+		if not targetpath then
 			local clean_path = nil
 			if basename ~= nil then
 				clean_path = basename
@@ -590,9 +590,7 @@ function pkgmgr.install_dir(type, path, basename, targetpath)
 			return nil, fgettext("Unable to install a mod as a $1", type)
 		end
 
-		if targetpath then
-			core.delete_dir(targetpath)
-		else
+		if not targetpath then
 			local targetfolder = basename
 			if targetfolder == nil then
 				targetfolder = pkgmgr.identify_modname(basefolder.path, "init.lua")
@@ -615,15 +613,15 @@ function pkgmgr.install_dir(type, path, basename, targetpath)
 			return nil, fgettext("Unable to install a game as a $1", type)
 		end
 
-		if targetpath then
-			core.delete_dir(targetpath)
-		else
+		if not targetpath then
 			targetpath = core.get_gamepath() .. DIR_DELIM .. basename
 		end
 	end
 
 	-- Copy it
-	core.delete_dir(targetpath)
+	if core.is_dir(targetpath) then
+		core.delete_dir(targetpath)
+	end
 	if not core.copy_dir(basefolder.path, targetpath, false) then
 		return nil,
 			fgettext("Failed to install $1 to $2", basename, targetpath)
